Add Readme and some cleanup
