--- 350-kubernetes-agent-application-pods/KubernetesMachine/admin.conf
@@ -1,19 +1,0 @@
-apiVersion: v1
-clusters:
-- cluster:
-    certificate-authority-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN5RENDQWJDZ0F3SUJBZ0lCQURBTkJna3Foa2lHOXcwQkFRc0ZBREFWTVJNd0VRWURWUVFERXdwcmRXSmwKY201bGRHVnpNQjRYRFRFNE1USXhNREU0TWpnd09Gb1hEVEk0TVRJd056RTRNamd3T0Zvd0ZURVRNQkVHQTFVRQpBeE1LYTNWaVpYSnVaWFJsY3pDQ0FTSXdEUVlKS29aSWh2Y05BUUVCQlFBRGdnRVBBRENDQVFvQ2dnRUJBTThpClV1Q2IxclFwcnVBN1pmV3I2b0wrd3ZFR3BCd0MxUEFwTHcxeitGcEhwd1lERUUvUHV5aWZaOHo3S2dDWlJ5NUYKN0RLQzlCVksvNjNhYzd6c1IvM042bEkzamt5V1ZReGFCeTQ2ZGthZTZFWnk5bmxUYk1vUlA0TTB2UGhWYzVXegpvR1ViNmlJaCtBL1Z1ZG9aZlI5Q0J1MWN1Z2FSUEdjU3M5VzhIZnpuaElOcmlTSTVvVHgyUWhDbE54QlBqOG1ZCll0bjBxUmxiV3pBamFBNmtTTEZsTXJMTmlqM1h0Y2NFYjl0NFhhT1hqQWtsSjRkTU1YS3I5OWJtOUYxZ0V3cjMKaVpJYWlMYXlteEk5TElHbGdPSkVYanlocDFSZlpOWC9VNmVOZ0ZkN2hOZWczNzdQaUNRRWxQampUeDZ0WnFpZwpQWHJyditNbXc5SVZuR3d0VzBNQ0F3RUFBYU1qTUNFd0RnWURWUjBQQVFIL0JBUURBZ0trTUE4R0ExVWRFd0VCCi93UUZNQU1CQWY4d0RRWUpLb1pJaHZjTkFRRUxCUUFEZ2dFQkFGQ0FTWW1ENTZ1emx5OEVaaHVlR1VKbzVaek0KMnR5UllSQkFwenBSdks5NHdzUGU3YUxreDBhY3BXRE1CcVUyL2NkNnZrWmp6MmdLcjdUTXRJN1lJVEx1dER4ZQpWeGxhdjNTTTJQY2FrTEo5V1J0QlVyK0tsWjY1ZzlYU24vMVlHbjVLeUFTYVpoL0J3RVQ3VERrZlJFQzNkWmdhCnFCRTZuZGM2aGkzSlZJSTVSM1d2WVNzRmR6K0o3NmlZQ1I0bEMzS25ramgrbmFEb1JlaGRkeDBzck9OOUZ3QTAKN1laSFZWV0NkQ1ROSUErdHVrSW53aWpNOURhZUFOUWcvbGdlY2k3ZkFYNXo0alZEd0ZzTEZ2eHhoMkRLc1VrZQpYUm4rMGpiMHhNK25vSm95clpqekRNajA3bXpMUnhjSEF4bFlyT3VsaDQ1a01TV1pZei9aWWtSNkY4dz0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
-    server: https://172.31.42.240:6443
-  name: kubernetes
-contexts:
-- context:
-    cluster: kubernetes
-    user: kubernetes-admin
-  name: kubernetes-admin@kubernetes
-current-context: kubernetes-admin@kubernetes
-kind: Config
-preferences: {}
-users:
-- name: kubernetes-admin
-  user:
-    client-certificate-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUM4akNDQWRxZ0F3SUJBZ0lJRXhISDJ3TXlwQmN3RFFZSktvWklodmNOQVFFTEJRQXdGVEVUTUJFR0ExVUUKQXhNS2EzVmlaWEp1WlhSbGN6QWVGdzB4T0RFeU1UQXhPREk0TURoYUZ3MHhPVEV5TVRBeE9ESTRNVEphTURReApGekFWQmdOVkJBb1REbk41YzNSbGJUcHRZWE4wWlhKek1Sa3dGd1lEVlFRREV4QnJkV0psY201bGRHVnpMV0ZrCmJXbHVNSUlCSWpBTkJna3Foa2lHOXcwQkFRRUZBQU9DQVE4QU1JSUJDZ0tDQVFFQXF3dXJBbWdjbDFWU20yNWQKNW5CK3BtbFYrZTV0dk9qZ295QU9DZjZ1cUxSdTR1WlRwOU1EVUJJRC8ydHlUZThtdlhDQ3Z3dkI4ZDFZTksvMApZbjFXbTZDRjdhbXRPamNqbHdtajFOWXFDLzJJK2pBd29xVkx1YmRpMlZ2Y2pvbG1sYW1FaWxVbnh0eVFRWEhSCnY1QVNXZWMyYWw4NVljWncxQjhRRzNxdFcxMy9wOTg1bWNabmFQTjA2SSsweERKdjd5cHJ5TDNJV3JBQ0NBejEKNlVGMjJyL2w4VkgvVy92K0M5RXNjcVlwL0VEc2RhTFpvcTlqNTlFNDhDNFVLWFdKeWpzSkwvc21oQng3NGlKaQp5dXQrN1RNWENyd05lVHIwN3d4ZFlXNk9HbkpLSWw3QkV2SllSbjRIbnd3UjdKYWZpZVQyNUgxbE15Ky9ydGtJClBUVzd5UUlEQVFBQm95Y3dKVEFPQmdOVkhROEJBZjhFQkFNQ0JhQXdFd1lEVlIwbEJBd3dDZ1lJS3dZQkJRVUgKQXdJd0RRWUpLb1pJaHZjTkFRRUxCUUFEZ2dFQkFGc0k5bngycEV6NlVnZ2ZFVmNEaGVnWjR3eVpxdDE2YjZPNQpBSzVKM3VaWUkrbnFEd0dHN0ZWRk04UzNEb2dBY1FpUzRFSlU0dWs0NG9qcU5qclBPd3ZIZ3VVd01EbXdmdkVUCkcwNU5nVnhvcXcyZ1lHRHd3ZWVweWF2UFM4ZDgxZlBIQnE0ak5Kay9jYjFtamg0TjNwMGVDTElOSTZuSjl2M3EKRUw3dzlRUGJEdTVDaGZ4MUQ5Vy9KdUk1RmhPOUJXTWJlVVg5S1hNV2FQbE9pcUQzSXorR0MyVGtmaFQ1K3RmbApIaHRkZkxGcWZjYU9DZDBMVXh4d3J4Y1dZRVM4bi82eFltYkwrUDV5S3RPenMyRExMTWxxeDhnZ0wraW5hdHZvCnhGQjczRXh5VTdVMm1SVXdTem56MlBlNXVVNG84NEdCNU04RUE1ckVObkhIREE1ZjVhTT0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
-    client-key-data: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFb2dJQkFBS0NBUUVBcXd1ckFtZ2NsMVZTbTI1ZDVuQitwbWxWK2U1dHZPamdveUFPQ2Y2dXFMUnU0dVpUCnA5TURVQklELzJ0eVRlOG12WENDdnd2QjhkMVlOSy8wWW4xV202Q0Y3YW10T2pjamx3bWoxTllxQy8ySStqQXcKb3FWTHViZGkyVnZjam9sbWxhbUVpbFVueHR5UVFYSFJ2NUFTV2VjMmFsODVZY1p3MUI4UUczcXRXMTMvcDk4NQptY1puYVBOMDZJKzB4REp2N3lwcnlMM0lXckFDQ0F6MTZVRjIyci9sOFZIL1cvditDOUVzY3FZcC9FRHNkYUxaCm9xOWo1OUU0OEM0VUtYV0p5anNKTC9zbWhCeDc0aUppeXV0KzdUTVhDcndOZVRyMDd3eGRZVzZPR25KS0lsN0IKRXZKWVJuNEhud3dSN0phZmllVDI1SDFsTXkrL3J0a0lQVFc3eVFJREFRQUJBb0lCQUhKWkVaRkJVdVBrK1ozSgprZUx0aiszbWU4RGpvZDlwL0V0eHNaRUMvM2J2Y1RZbGJ3WENMdlBIcDM2Sk1WaGlxZzU1ZWg4SVdEYytsNnBaCmUxS2dvbGJXWU5LNjNZRU92YTlLbFlsRFI4Yk9UMVdiQTlpb0V0RHNoWnYwS3FYcnUzWVFadmduNTdpdFdZODIKN1ZkWWxMd2ZzNlEzVnVCRHpzZ3dvQThPMjNnNWJkT2pBVXQ2bHhkb25mSGordE5abTNML1ZHam5IR3lnRGVLcgpsWHVYTUwybTV4UmhNQ1UzWk9YSGNsc01zZU56Y1lrNWpYUm8rR213TlFXeDdpaXM4bXZIWnFvTUJUTC8rK1VhCko5RUxibzhYWnlGV2NudDJlTklKYXpkaXMzSnh3WGpWREh2c2VvVk1nQ0NHcWllMGplNytwRnhPSlRudzF3NDAKSXpROG9WVUNnWUVBeE9xdlpKVjlWUm93SXlEYlhyWmZvQWR4V1R1ZjFaUVJzdlNtTkZzWDZKSGxyVERPUEpqNQo4MlQvL3ZwVWQ4dHVwZjc4U0NqbTUwSXF6Qlo4U1RKcTNrS2tjYkFXdVordFRkaXdGdUM1UmZGWm9HMERSTjcyClQ5SFFRWXhabFpHdGcwN1BkWXFPbzZlMGYzWkt1K3l6eUhhaFZEbDNSNkpQYWt2QnlkVm90dHNDZ1lFQTNsM00KOVA0anBrYXppUkJRNXNFYnB2NmorZy9TZW1IUDJVeThkR1cxQzAxNThTcm5HY1NKSXlVQU5vMlFaRmZuUTNCUQorYzhOdWtmN29yUzgyc0ZRaGRDSWttSmpERUtyVzQ4djJ5azNyU3RJaFMyc085NUg5ay9GQnM5eG90ejFuL0FOCmx3ZlFiSVpNalpyUDVJc244cFM1ZWFyOTRTd0pnOSswSWRZM255c0NnWUFzN0RrOUlobG9tOFBWajc1YitoRVEKcWdJQzBHWmpUQmRKa3lTRGx6QjZIeWFkZngvZGxTeVlpVGNaalRSTjYvdi9KSStiU3JGNm1vNjI1bjVPR0pGbQp5UjlJekZveTJLbXBzOEJnUjc0Y3BiWERzbnJuY0ZJUzNaTzhFRyt6Zm50dUZFR2gyQTA0YXA3ejRTR2xnNWQwCktSUWU2Y2s3YjF6Y3JCRzJvTmYxdHdLQmdHRk4zcEFlWFc4M1pQekhYUDk4NVNMUUx1VVA0MlNXUktPc0JwTlEKcGthSjRxQjd5NkRrOWJXaFRWU1I0ZXRtNDlZSTRnWk9hbzFuVGwvR2d4V2VuaVBRT3RNRHBxQUkxSDgrU3d5MgpIcHp5WWN1VlR0YnJ6OFdpNW02eWUrRnVtelVma0hMVWdvbmF3c2o4UTVpWXBDUnNuRFQzZEIwQ2JpOTh4cEEvCjh4cjlBb0dBQVk0Y3QyNnoxSGRNVHBqZmowWUd3MUsybHVYcCtlSmZjUVE1bGVraWZ5VS85S1l2WDJMeHlseUIKOXNsTDJiVWFhcnJTVjRCSXRsSkNwTGxseEFsVzJZQStVbjJOeUFQT2dsM0FaNlZYU2Vja0s5b2JFNmJNdzh5RwpwbnRscGZBNU1Tbi9raGVKK3NIckpUcGZPVDBKUERoVFAzWWNxUVBqN1R0dHJKZ0R1UEk9Ci0tLS0tRU5EIFJTQSBQUklWQVRFIEtFWS0tLS0tCg==
