+++ 350-kubernetes-agent-application-pods/KubernetesMachine/ctmContainer/c360srv01-job.yaml
@@ -10,12 +10,12 @@ spec:
       containers:
       - name: c360srv01
         image: joegoldberg/controlm:appimage
-        imagePullPolicy: Always
         env:
         - name: LOOPCTR
           value: "10"
         - name: STIME
           value: "5"
+      restartPolicy: Never
       imagePullSecrets:
       - name: regcred
-      restartPolicy: Never
+
